+++ src/client/visualizations/line-chart/base-chart/foreground/foreground.tsx
@@ -17,10 +17,10 @@
 import { Timezone } from "chronoshift";
 import * as React from "react";
 import { ReactNode } from "react";
-import { Stage } from "../../../../common/models/stage/stage";
-import { Nullary } from "../../../../common/utils/functional/functional";
-import { Interaction, isHighlight, isHover } from "../interactions/interaction";
-import { ContinuousScale } from "../utils/continuous-types";
+import { Stage } from "../../../../../common/models/stage/stage";
+import { Nullary } from "../../../../../common/utils/functional/functional";
+import { Interaction, isHighlight, isHover } from "../../interactions/interaction";
+import { ContinuousScale } from "../../utils/continuous-types";
 import { HighlightModal } from "./highlight-modal";
 import { HoverTooltip } from "./hover-tooltip";
 import { SelectionOverlay } from "./selection-overlay";
