+++ src/client/visualizations/line-chart/base-chart/foreground/highlight-modal.tsx
@@ -16,12 +16,12 @@
 
 import { Timezone } from "chronoshift";
 import * as React from "react";
-import { formatValue } from "../../../../common/utils/formatter/formatter";
-import { Nullary } from "../../../../common/utils/functional/functional";
-import { HighlightModal as BaseHighlightModal } from "../../../components/highlight-modal/highlight-modal";
-import { toPlywoodRange } from "../interactions/highlight-clause";
-import { Highlight } from "../interactions/interaction";
-import { ContinuousScale } from "../utils/continuous-types";
+import { formatValue } from "../../../../../common/utils/formatter/formatter";
+import { Nullary } from "../../../../../common/utils/functional/functional";
+import { HighlightModal as BaseHighlightModal } from "../../../../components/highlight-modal/highlight-modal";
+import { toPlywoodRange } from "../../interactions/highlight-clause";
+import { Highlight } from "../../interactions/interaction";
+import { ContinuousScale } from "../../utils/continuous-types";
 
 interface HighlightModalProps {
   interaction: Highlight;
