+++ src/client/visualizations/line-chart/base-chart/foreground/hover-guide.tsx
@@ -16,9 +16,9 @@
 
 import * as d3 from "d3";
 import * as React from "react";
-import { Stage } from "../../../../common/models/stage/stage";
-import { Hover } from "../interactions/interaction";
-import { ContinuousScale } from "../utils/continuous-types";
+import { Stage } from "../../../../../common/models/stage/stage";
+import { Hover } from "../../interactions/interaction";
+import { ContinuousScale } from "../../utils/continuous-types";
 
 interface HoverGuideProps {
   hover: Hover;
