+++ src/client/visualizations/line-chart/base-chart/foreground/selection-overlay.tsx
@@ -17,12 +17,12 @@
 import { Timezone } from "chronoshift";
 import { Range } from "plywood";
 import * as React from "react";
-import { Stage } from "../../../../common/models/stage/stage";
-import { Highlighter } from "../../../components/highlighter/highlighter";
-import { constructRange } from "../interactions/continuous-range";
-import { Interaction, isDragging, isHighlight } from "../interactions/interaction";
-import { ContinuousRange, ContinuousScale } from "../utils/continuous-types";
-import { isValidClause } from "../utils/is-valid-clause";
+import { Stage } from "../../../../../common/models/stage/stage";
+import { Highlighter } from "../../../../components/highlighter/highlighter";
+import { constructRange } from "../../interactions/continuous-range";
+import { Interaction, isDragging, isHighlight } from "../../interactions/interaction";
+import { ContinuousRange, ContinuousScale } from "../../utils/continuous-types";
+import { isValidClause } from "../../utils/is-valid-clause";
 
 interface SelectionOverlayProps {
   interaction: Interaction;
